+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2016-04-26
+### Added
+- Flamegraph
+- Dotnetcli
+- F5 cloud service
+
 ## 2016-03-29
 ### Added
 - Jboss